Conversions between uint128 and uint256 (#220)
* expose uint256 instead of uint128 in Vault, use uint256 in pools

* add changes suggested in PR review

+++ contracts/vault/interfaces/IPoolQuote.sol
@@ -20,15 +20,15 @@ import "./IPoolQuoteStructs.sol";
 interface IPoolQuote {
     function quoteOutGivenIn(
         IPoolQuoteStructs.QuoteRequestGivenIn calldata request,
-        uint128[] calldata balances,
+        uint256[] calldata balances,
         uint256 indexIn,
         uint256 indexOut
-    ) external returns (uint128 amountOut);
+    ) external returns (uint256 amountOut);
 
     function quoteInGivenOut(
         IPoolQuoteStructs.QuoteRequestGivenOut calldata request,
-        uint128[] calldata balances,
+        uint256[] calldata balances,
         uint256 indexIn,
         uint256 indexOut
-    ) external returns (uint128 amountIn);
+    ) external returns (uint256 amountIn);
 }
